Unify READMEs
Unify the kvm-ioctls and kvm-bindings READMEs.

While doing this, I dropped parts of the kvm-bindings readme that
covered how to add the crate as a dependency in `Cargo.toml`. My
reasoning for this is that the steps were fairly outdated (predating
rust edition 2018, due to advising for use of `extern crate`), and also
should be fairly obvious for anyone who has used Rust before.

Note that this version of the README assumes that renames of the
repository and CI pipelines have already taken place.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>